feat: 新增下载管理功能，包含下载服务、界面及通知支持

+++ core/data/src/main/java/com/imcys/bilibilias/data/model/download/DownloadSubTask.kt
@@ -1,6 +1,5 @@
 package com.imcys.bilibilias.data.model.download
 
-import com.imcys.bilibilias.database.entity.download.DownloadStage
 import com.imcys.bilibilias.database.entity.download.DownloadState
 import com.imcys.bilibilias.database.entity.download.DownloadSubTaskType
 import java.util.Date
@@ -9,7 +8,7 @@ data class DownloadSubTask(
     val segmentId: Long,
     val downloadUrl: String,
     val savePath: String,
-    val progress: Double = 0.0,
+    val progress: Float = 0.0f,
     val subTaskType: DownloadSubTaskType,
     val downloadState: DownloadState = DownloadState.WAITING,
     val createTime: Date = Date(),
